+++ curves/curve25519/.gitignore
@@ -1,0 +1,1 @@
+/farf/
