+++ curves/curve25519/src/curve_syscall_traits.rs
