+++ programs/sbf/Cargo.toml
@@ -31,6 +31,7 @@ solana-accounts-db = { path = "../../accounts-db", version = "=2.0.0" }
 solana-bpf-loader-program = { path = "../bpf_loader", version = "=2.0.0" }
 solana-cli-output = { path = "../../cli-output", version = "=2.0.0" }
 solana-compute-budget = { path = "../../compute-budget", version = "=2.0.0" }
+solana-curve25519 = { path = "../../curves/curve25519", version = "=2.0.0" }
 solana-ledger = { path = "../../ledger", version = "=2.0.0" }
 solana-logger = { path = "../../logger", version = "=2.0.0" }
 solana-measure = { path = "../../measure", version = "=2.0.0" }
@@ -218,5 +219,6 @@ members = [
 #
 # There is a similar override in `../../Cargo.toml`.  Please keep both comments
 # and the overrides in sync.
+solana-curve25519 = { path = "../../curves/curve25519" }
 solana-program = { path = "../../sdk/program" }
 solana-zk-token-sdk = { path = "../../zk-token-sdk" }
