+++ programs/sbf/rust/curve25519/src/lib.rs
@@ -2,8 +2,8 @@
 
 extern crate solana_program;
 use {
+    solana_curve25519::{edwards, ristretto, scalar},
     solana_program::{custom_heap_default, custom_panic_default, msg},
-    solana_zk_token_sdk::curve25519::{edwards, ristretto, scalar},
 };
 
 #[no_mangle]
