+++ zk-token-sdk/Cargo.toml
@@ -14,6 +14,7 @@ base64 = { workspace = true }
 bytemuck = { workspace = true, features = ["derive"] }
 num-derive = { workspace = true }
 num-traits = { workspace = true }
+solana-curve25519 = { workspace = true }
 solana-program = { workspace = true }
 thiserror = { workspace = true }
 
@@ -25,7 +26,6 @@ aes-gcm-siv = { workspace = true }
 bincode = { workspace = true }
 byteorder = { workspace = true }
 curve25519-dalek = { workspace = true, features = ["serde"] }
-getrandom = { version = "0.1", features = ["dummy"] }
 itertools = { workspace = true }
 lazy_static = { workspace = true }
 merlin = { workspace = true }
