+++ zk-token-sdk/src/errors.rs
@@ -3,6 +3,7 @@
 use crate::range_proof::errors::RangeProofGenerationError;
 use {
     crate::{range_proof::errors::RangeProofVerificationError, sigma_proofs::errors::*},
+    solana_curve25519::errors::ElGamalError,
     thiserror::Error,
 };
 
@@ -16,24 +17,6 @@ pub enum AuthenticatedEncryptionError {
     SeedLengthTooLong,
     #[error("failed to deserialize")]
     Deserialization,
-}
-
-#[derive(Error, Clone, Debug, Eq, PartialEq)]
-pub enum ElGamalError {
-    #[error("key derivation method not supported")]
-    DerivationMethodNotSupported,
-    #[error("seed length too short for derivation")]
-    SeedLengthTooShort,
-    #[error("seed length too long for derivation")]
-    SeedLengthTooLong,
-    #[error("failed to deserialize ciphertext")]
-    CiphertextDeserialization,
-    #[error("failed to deserialize public key")]
-    PubkeyDeserialization,
-    #[error("failed to deserialize keypair")]
-    KeypairDeserialization,
-    #[error("failed to deserialize secret key")]
-    SecretKeyDeserialization,
 }
 
 #[cfg(not(target_os = "solana"))]
