+++ zk-token-sdk/src/zk_token_elgamal/pod/grouped_elgamal.rs
@@ -3,14 +3,12 @@
 #[cfg(not(target_os = "solana"))]
 use crate::encryption::grouped_elgamal::GroupedElGamalCiphertext;
 use {
-    crate::{
-        errors::ElGamalError,
-        zk_token_elgamal::pod::{
-            elgamal::{ElGamalCiphertext, DECRYPT_HANDLE_LEN, ELGAMAL_CIPHERTEXT_LEN},
-            pedersen::{PedersenCommitment, PEDERSEN_COMMITMENT_LEN},
-            Pod, Zeroable,
-        },
+    crate::zk_token_elgamal::pod::{
+        elgamal::{ElGamalCiphertext, DECRYPT_HANDLE_LEN, ELGAMAL_CIPHERTEXT_LEN},
+        pedersen::{PedersenCommitment, PEDERSEN_COMMITMENT_LEN},
+        Pod, Zeroable,
     },
+    solana_curve25519::errors::ElGamalError,
     std::fmt,
 };
